+++ src/ScottPlot5/ScottPlot5/TickGenerators/NumericAutomatic.cs
@@ -23,18 +23,19 @@ public class NumericAutomatic : ITickGenerator
 
     public static string DefaultLabelFormatter(double value)
     {
-        CultureInfo culture = CultureInfo.InvariantCulture;
+        // use system culture
+        // https://github.com/ScottPlot/ScottPlot/issues/3688
 
         // if the number is round or large, use the numeric format
         // https://docs.microsoft.com/en-us/dotnet/standard/base-types/standard-numeric-format-strings#the-numeric-n-format-specifier
         bool isRoundNumber = (int)value == value;
         bool isLargeNumber = Math.Abs(value) > 1000;
         if (isRoundNumber || isLargeNumber)
-            return value.ToString("N0", culture);
+            return value.ToString("N0");
 
         // otherwise the number is probably small or very precise to use the general format (with slight rounding)
         // https://docs.microsoft.com/en-us/dotnet/standard/base-types/standard-numeric-format-strings#the-general-g-format-specifier
-        string label = Math.Round(value, 10).ToString("G", culture);
+        string label = Math.Round(value, 10).ToString("G");
         return label == "-0" ? "0" : label;
     }